Move generic middleware to client/controller
