+++ client/my-sites/themes/controller.js
@@ -1,11 +1,9 @@
 /**
  * External Dependencies
  */
-import ReactDom from 'react-dom';
 import React from 'react';
 import { Provider as ReduxProvider } from 'react-redux';
 import omit from 'lodash/omit';
-import startsWith from 'lodash/startsWith';
 
 /**
  * Internal Dependencies
@@ -133,35 +131,4 @@ export function details( context, next ) {
 	context.primary = makeElement( ThemeSheetComponent, Head, context.store, props );
 	context.secondary = null; // When we're logged in, we need to remove the sidebar.
 	next();
-}
-
-// Generic middleware -- move to client/controller.js?
-// lib/react-helpers isn't probably middleware-specific enough
-export function renderElements( context ) {
-	renderPrimary( context );
-	renderSecondary( context );
-}
-
-function renderPrimary( context ) {
-	const { path } = context.params;
-	// FIXME: temporary hack until we have a proper isomorphic, one tree routing solution. Do NOT do this!
-	const sheetsDomElement = startsWith( path, '/themes' ) && document.getElementsByClassName( 'themes__sheet' )[0];
-	const mainDomElement = startsWith( path, '/design' ) && document.getElementsByClassName( 'themes main' )[0];
-	if ( ! sheetsDomElement && ! mainDomElement ) {
-		ReactDom.render(
-			context.primary,
-			document.getElementById( 'primary' )
-		);
-	}
-}
-
-function renderSecondary( context ) {
-	if ( context.secondary === null ) {
-		ReactDom.unmountComponentAtNode( document.getElementById( 'secondary' ) );
-	} else if ( context.secondary !== undefined ) {
-		ReactDom.render(
-			context.secondary,
-			document.getElementById( 'secondary' )
-		);
-	}
 }
